# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para conversion de PDF, IIIF Presentation v2 (con v3 compatible), IIIF Image API, administracion y migracion. Las rutas recomendadas usan /api/v1 y los endpoints legacy se mantienen por compatibilidad temporal.",
        "title": "IIIF PDF Server API",
        "contact": {},
        "version": "2.1"
    },
    "basePath": "/",
    "paths": {
        "/api/iiif/v3/{id}/manifest": {
            "get": {
                "description": "Endpoint de compatibilidad para clientes IIIF Presentation v3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/iiif/{id}/manifest": {
            "get": {
                "description": "Devuelve un manifiesto v2 con secuencia, canvases y rangos jerarquicos derivados de los marcadores PDF. El parametro pages permite seleccionar paginas y poda rangos vacios.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifestV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/documents": {
            "get": {
                "description": "Lista documentos con filtros opcionales por proyecto y tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Listar documentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proyecto",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PDFDocument"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/upload": {
            "post": {
                "description": "Recibe archivo PDF y crea documento con paginas convertidas.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Subir y convertir PDF",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Archivo PDF",
                        "name": "pdf",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Proyecto",
                        "name": "project",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1241,
                        "description": "Ancho maximo de imagen",
                        "name": "max_width",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1754,
                        "description": "Alto maximo de imagen",
                        "name": "max_height",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 150,
                        "description": "Resolucion de render",
                        "name": "dpi",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "jpg",
                        "description": "Formato de salida (jpg o png)",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 85,
                        "description": "Calidad JPEG",
                        "name": "quality",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFDocument"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Obtener documento por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFDocument"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Eliminar documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.okMessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/iiif/{id}/manifest": {
            "get": {
                "description": "Devuelve un manifiesto v2 con secuencia, canvases y rangos jerarquicos derivados de los marcadores PDF. El parametro pages permite seleccionar paginas y poda rangos vacios.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifestV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/iiif/{id}/manifest/v3": {
            "get": {
                "description": "Endpoint de compatibilidad para clientes IIIF Presentation v3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/iiif/2/{identifier}/info.json": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener info IIIF Image API v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador IIIF, por ejemplo documento_page_1",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFImageInfoV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "detalle interno"
                },
                "error": {
                    "type": "string",
                    "example": "login requerido"
                }
            }
        },
...
        "handlers.okMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
        "models.IIIFImageInfoV2": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "string"
                },
                "@id": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "profile": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "sizes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFSize"
                    }
                },
                "tiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFTile"
                    }
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "models.IIIFManifest": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFCanvas"
                    }
                },
                "label": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "project_iiif_session",
            "in": "cookie"
        }
    }
}